move Fact to new fact.py
