+++ src/hamster/lib/dbus.py
@@ -3,7 +3,7 @@ import dbus
 import dbus.service
 from calendar import timegm
 from dbus.mainloop.glib import DBusGMainLoop as DBusMainLoop
-from hamster.lib import Fact
+from hamster.lib.fact import Fact
 
 
 """D-Bus communication utilities."""
